split rocsparse_handle.cpp
